# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BCA Email Import Backend
  version: 1.0.0
  description: Owner API and worker-only BCA receipt ingestion. Monetary values are exact decimal strings in IDR major units.
servers:
  - url: http://127.0.0.1:18080
paths:
...
  /auth/google/callback:
    get:
      operationId: googleCallback
      description: Google redirects here; a successful HTML page shows a one-time application login code.
      parameters:
        - {name: state, in: query, required: true, schema: {type: string}}
        - {name: code, in: query, required: true, schema: {type: string}}
      responses:
        '200': {description: One-time code in HTML}
        '400': {description: Invalid or expired authorization state}
        '403': {description: Google subject is not the configured owner}
  /api/v1/auth/exchange:
    post:
      operationId: exchangeCode
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [code], additionalProperties: false, properties: {code: {type: string}}}
      responses:
        '200': {description: Application tokens, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
        '401': {description: Invalid code, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  /api/v1/auth/refresh:
    post:
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema: {type: object, required: [refresh_token], additionalProperties: false, properties: {refresh_token: {type: string}}}
      responses:
        '200': {description: Rotated application tokens, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}
        '401': {description: Invalid or reused refresh token, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  /api/v1/auth/logout:
    post:
      operationId: logout
      security: [{ownerBearer: []}]
      responses:
        '204': {description: Session family revoked}
        '401': {description: Unauthorized}
  /api/v1/me:
    get:
      operationId: me
      security: [{ownerBearer: []}]
      responses:
        '200':
          description: Owner profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: string, format: uuid}
                  email: {type: string, format: email}
                  timezone: {type: string}
        '401': {description: Unauthorized}
  /api/v1/integrations/gmail:
    get:
      operationId: gmailStatus
      security: [{ownerBearer: []}]
      responses:
        '200': {description: Connection status and last successful discovery}
    delete:
      operationId: gmailDisconnect
      security: [{ownerBearer: []}]
      responses:
        '204': {description: Credentials removed; transactions retained}
        '404': {description: Integration missing}
  /api/v1/integrations/gmail/sync:
    post:
      operationId: gmailSync
      security: [{ownerBearer: []}]
      responses:
        '202': {description: Discovery scheduled, content: {application/json: {schema: {type: object, properties: {operation_id: {type: string, format: uuid}, status: {type: string}}}}}}
        '409': {description: Integration unavailable}
  /api/v1/integrations/gmail/sync/{id}:
    get:
      operationId: gmailSyncOutcome
      security: [{ownerBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200':
          description: Durable discovery operation status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: string, format: uuid}
                  status: {type: string, enum: [queued, running, completed, failed]}
                  reason_code: {type: string, nullable: true}
                  created_at: {type: string, format: date-time}
                  finished_at: {type: string, format: date-time, nullable: true}
        '404': {description: Operation not found}
  /api/v1/imports:
    get:
      operationId: listImports
      security: [{ownerBearer: []}]
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: state, in: query, schema: {type: string, enum: [queued, processing, retry_wait, completed, ignored, unsupported, needs_review, failed]}, description: Filter jobs by state}
      responses:
        '200': {description: BCA processing jobs without email content, content: {application/json: {schema: {$ref: '#/components/schemas/Page'}}}}
  /api/v1/imports/{id}:
    get:
      operationId: getImport
      security: [{ownerBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Safe import outcome}
        '404': {description: Import not found}
  /api/v1/imports/{id}/source:
    get:
      operationId: findImportSource
      description: Owner-only metadata lookup. Fetches the From header first, then subject, date and Message-ID for an exact BCA sender. No message body is fetched. Paste gmail_search into Gmail to open the email.
      security: [{ownerBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200':
          description: Gmail source metadata
          content:
            application/json:
              schema:
                type: object
                required: [id, subject, date, gmail_search]
                properties:
                  id: {type: string, format: uuid}
                  subject: {type: string}
                  date: {type: string}
                  gmail_search: {type: string, description: Paste into the connected Gmail account search bar; empty if the message lacks a usable Message-ID header}
        '404': {description: Import not found or not owned}
        '503': {description: Gmail source unavailable}
  /api/v1/imports/{id}/inspect:
    get:
      operationId: inspectUnsupportedImport
      description: Owner-only on-demand inspection of one unsupported job. Refetches the BCA email after sender and receiver-authentication checks, returns bounded layout clues, and does not store the message or change the job.
      security: [{ownerBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200':
          description: Current parser diagnosis
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: {type: string, format: uuid}
                  job_state: {type: string, enum: [unsupported]}
                  diagnostic:
                    type: object
                    properties:
                      parser_outcome: {type: string, enum: [supported, unsupported, ignored, needs_review, not_checked]}
                      reason_code: {type: string}
                      retry_recommended: {type: boolean, description: True when the current parser supports this previously unsupported job; inspection itself does not retry it}
                      subject: {type: string, description: Bounded subject preview; this field contains email content}
                      transaction_type: {type: string, description: Bounded transaction type preview}
                      transfer_type: {type: string, description: Bounded transfer type preview}
                      detected_fields: {type: array, items: {type: string}, description: Allowlisted field labels only}
                      receipt_kind: {type: string}
        '404': {description: Import not found or not owned}
        '409': {description: Job is not unsupported}
        '503': {description: Gmail inspection unavailable}
  /api/v1/imports/{id}/retry:
    post:
      operationId: retryImport
      security: [{ownerBearer: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '202': {description: Retry queued}
        '409': {description: Job not retryable}
  /api/v1/webhooks/bca:
    post:
      operationId: ingestBCA
      description: Worker-only. Source job and matching persisted delivery payload are required.
      security: [{workerBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/BCAEvent'}
      responses:
        '201': {description: New financial event, content: {application/json: {schema: {$ref: '#/components/schemas/IngestResult'}}}}
        '200': {description: Identical duplicate, content: {application/json: {schema: {$ref: '#/components/schemas/IngestResult'}}}}
        '401': {description: Worker token missing or invalid, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '409': {description: Contradictory duplicate or disconnected integration, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
        '422': {description: Invalid event or unknown source, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  /api/v1/transactions:
    get:
      operationId: listTransactions
      security: [{ownerBearer: []}]
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Cursor'}
        - {name: from, in: query, schema: {type: string, format: date}}
        - {name: to, in: query, schema: {type: string, format: date}}
        - {name: bank, in: query, schema: {type: string}}
        - {name: classification, in: query, schema: {type: string, enum: [unclassified, expense, internal_transfer]}}
        - {name: account, in: query, schema: {type: string}}
      responses:
        '200': {description: Transaction page, content: {application/json: {schema: {$ref: '#/components/schemas/Page'}}}}
...
  /api/v1/owned-accounts:
    get:
      operationId: listOwnedAccounts
      security: [{ownerBearer: []}]
      responses:
        '200': {description: Masked owned accounts}
    post:
      operationId: addOwnedAccount
      security: [{ownerBearer: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bank, account_number, label]
              additionalProperties: false
              properties:
                bank: {type: string}
                account_number: {type: string, pattern: '^[0-9]{5,40}$', writeOnly: true}
                label: {type: string}
      responses:
        '201': {description: Only masked identifier is returned}
components:
  securitySchemes:
    ownerBearer: {type: http, scheme: bearer, bearerFormat: JWT}
    workerBearer: {type: http, scheme: bearer}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 50}}
    Cursor: {name: cursor, in: query, schema: {type: string}}
  schemas:
    Error:
      type: object
      required: [code, request_id]
      properties:
        code: {type: string}
        request_id: {type: string}
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        expires_in: {type: integer, example: 900}
...
    IngestResult:
      type: object
      required: [id, created]
      properties:
        id: {type: string, format: uuid}
        created: {type: boolean}
...
    Page:
      type: object
      required: [items, next_cursor]
      properties:
        items: {type: array, items: {type: object}}
        next_cursor: {type: string}